Add orientation and noise
Entities now have an orientation which is taken into consideration by
the code that emits blips, so that blips are positioned as if the sensor
was on the entity with the brain itself.

There will also be a very small amount of noise on the delta of the
blips to simulate slight information loss.